Double Coordinate System -Coordinate system is no longer int based -Coordinate system is now double based -Allows for more variation in movement speeds -First Step in vector based movement

+++ src/core/player/Player.cpp
@@ -4,7 +4,7 @@
 
 #include "../../Util.h"
 
-Player::Player(int x, int y, Sprite sprite) :
+Player::Player(double x, double y, Sprite sprite) :
     Agent(x, y, true, sprite)
 {
     action = new ProjectileTestAction(this);
